use new parser and better error message printing
